+++ Services/Kafka/V2/Model/ShowPartitionMessageRespMessage.cs
@@ -43,7 +43,7 @@ namespace HuaweiCloud.SDK.Kafka.V2.Model
         /// 消息位置。
         /// </summary>
         [JsonProperty("message_offset", NullValueHandling = NullValueHandling.Ignore)]
-        public int? MessageOffset { get; set; }
+        public long? MessageOffset { get; set; }
 
         /// <summary>
         /// 消息大小，单位字节。
